requirements.txt
-i https://pypi.org/simple
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
entrypoints==0.4
flatbuffers==23.3.3
gast==0.4.0
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
libclang==15.0.6.1
markdown==3.4.1
markdown-it-py==2.2.0
markupsafe==2.1.2
mdurl==0.1.2
mouseinfo==0.1.3
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pillow==9.4.0
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyautogui==0.9.53
pydeck==0.8.0
pygetwindow==0.0.9
pygments==2.14.0
pympler==1.0.1
pymsgbox==1.0.9
pyperclip==1.8.2
pyrect==0.2.0
pyrsistent==0.19.3
pyscreeze==0.1.28
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
scipy==1.10.1
semver==2.13.0
setuptools==67.6.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
streamlit==1.20.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
zipp==3.15.0
